ci: Run deploy workflow weekly
Currently the deploy workflow is only run on push to this repo or when
manually triggered.
This leads to a quite outdated deploy.

By running this weekly (arbitrarily chosen 5AM Sunday) we get a
reasonably up-to-date deploy for our api-docs.

In the future we might consider a different approach but for now this
should be sufficient.